post currencies in progress

+++ src/main/java/org/kivislime/currencyexchange/model/Currency.java
@@ -1,19 +1,19 @@
 package org.kivislime.currencyexchange.model;
 
 public class Currency {
-    private final int id;
+    private final Long id;
     private final String code;
     private final String fullName;
     private final String sign;
 
-    public Currency(int id, String code, String fullName, String sign) {
+    public Currency(Long id, String code, String fullName, String sign) {
         this.id = id;
         this.code = code;
         this.fullName = fullName;
         this.sign = sign;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
